feat: update package revision page header (#5)
This change updates the Package Revision Page header to show package name and revision as the title to clarify that single revision a the package is being viewed. The package descriptor is moved to the right of the header and shown with the lifecycle of the package which is now a standard div. The chip was removed to reduce confusion on whether it was a clickable element.