feat: create new revisions of a package (#8)
This change allows for new revisions of a package to be created. When a user is on the latest published revision of a package on the Package Revision Page, a 'Create New Revision' button will appear. Clicking on this button will copy the package into a new draft revision and update the page to show the new revision. When a new revision of a package is published, the sync (if one is already created for the package) will automatically update to the new published revision.