feat: add restore package revision (#22)
This change allows previously published revisions of a package to be restored. This is completed by clicking the new 'Restore Revision' button on the Package Revision Page when a previously published package revision is shown. Restore Revision will create a new revision of the package, updating the resources of the new revision to the resources of the revision the restore is requested for.