feat: allow packages to upgrade to latest blueprint (#54)
This change allows packages to be upgraded to use the latest published blueprint revision. A package can be upgraded by clicking the new 'Upgrade to Latest Blueprint' button on the Package Revision Page. This button shows when the latest published revision of a package is shown, and a newer published revision exists for the package's blueprint. Additionally, the Packages Table shows a visual indicator when an upgrade is available for a package.